Fix cppcheck warnings

+++ rigs/lowe/lowe.c
@@ -114,7 +114,7 @@ int lowe_get_freq(RIG *rig, vfo_t vfo, freq_t *freq)
 {
     char freqbuf[16];
     int freq_len, retval;
-    float f_freq;
+    double f_freq;
 
     retval = lowe_transaction(rig, "FRQ?" EOM, 5, freqbuf, &freq_len);
 
@@ -125,7 +125,7 @@ int lowe_get_freq(RIG *rig, vfo_t vfo, freq_t *freq)
 
     freqbuf[freq_len < 16 ? freq_len : 15] = '\0';
 
-    sscanf(freqbuf + 1, "%f", &f_freq);
+    sscanf(freqbuf + 1, "%lf", &f_freq);
     *freq = f_freq * 1000;
 
     return retval;
